+++ src/test/browser.js
@@ -1,6 +1,5 @@
 require('tap-browser-color')();
 require('babel-polyfill');
 
-require('./');
-require('../experimental/test');
+require('./index');
 require('./react');