+++ src/test/imports-spec.js
@@ -1,5 +1,5 @@
 import test from 'tape-catch';
-import 'luma.gl/headless';
+import 'luma.gl';
 
 import DeckGL from '../../react';