+++ src/test/index.js
@@ -18,8 +18,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import 'babel-polyfill';
-
 import './imports-spec';
 import './lib';
 import './layers';
+
+import '../experimental/test';