+++ src/test/lib/layer.spec.js
@@ -1,5 +1,4 @@
 import test from 'tape-catch';
-import 'luma.gl/headless';
 import {Layer} from '../..';
 
 const dataVariants = [